+++ mark1/lib/configs/themes/theme_collections.dart
@@ -1,7 +1,61 @@
 // Theme collections
 
-// template for themes
-// sixPM theme
+// * Template for themes *
+// LukeCreated Original Theme ('original)
+// https://rydmike.com/flexcolorscheme/themesplayground-latest/
+// // Theme config for FlexColorScheme version 7.3.x. Make sure you use
+// // same or higher package version, but still same major version. If you
+// // use a lower package version, some properties may not be supported.
+// // In that case remove them after copying this theme to your app.
+// theme: FlexThemeData.light(
+//   colors: const FlexSchemeColor(
+//     primary: Color(0xff1565c0),
+//    primaryContainer: Color(0xff80d8ff),
+//    secondary: Color(0xffffca28),
+//    secondaryContainer: Color(0xffe4eaff),
+//    tertiary: Color(0xffea80fc),
+//    tertiaryContainer: Color(0xffcbd6ff),
+//    appBarColor: Color(0xffe4eaff),
+//    error: Color(0xffb00020),
+//   ),
+//   surfaceMode: FlexSurfaceMode.levelSurfacesLowScaffold,
+//   blendLevel: 7,
+//   subThemesData: const FlexSubThemesData(
+//     blendOnLevel: 10,
+//     blendOnColors: false,
+//     useTextTheme: true,
+//     useM2StyleDividerInM3: true,
+//     alignedDropdown: true,
+//     useInputDecoratorThemeInDialogs: true,
+//   ),
+//   visualDensity: FlexColorScheme.comfortablePlatformDensity,
+//   useMaterial3: true,
+//   swapLegacyOnMaterial3: true,
+//   // To use the Playground font, add GoogleFonts package and uncomment
+//   // fontFamily: GoogleFonts.notoSans().fontFamily,
+// ),
+// darkTheme: FlexThemeData.dark(
+//   primary: Color(0xff00bcd4),
+//    primaryContainer: Color(0xff3f51b5),
+//    secondary: Color(0xffd6a400),
+//    secondaryContainer: Color(0xffc9d8f2),
+//    tertiary: Color(0xffc51162),
+//    tertiaryContainer: Color(0xFF4C74CF),
+//    appBarColor: Color(0xffc9d8f2),
+//    error: Color(0xffcf6679),
+//   ),
+//   surfaceMode: FlexSurfaceMode.levelSurfacesLowScaffold,
+//   visualDensity: FlexColorScheme.comfortablePlatformDensity,
+//   useMaterial3: true,
+//   swapLegacyOnMaterial3: true,
+//   // To use the Playground font, add GoogleFonts package and uncomment
+//   // fontFamily: GoogleFonts.notoSans().fontFamily,
+// ),
+// // If you do not have a themeMode switch, uncomment this line
+// // to let the device system mode control the theme mode:
+// // themeMode: ThemeMode.system,
+
+// seaTheSky theme
 // https://rydmike.com/flexcolorscheme/themesplayground-latest/
 // // Theme config for FlexColorScheme version 7.3.x. Make sure you use
 // // same or higher package version, but still same major version. If you
@@ -46,14 +100,6 @@
 //     error: Color(0xffcf6679),
 //   ),
 //   surfaceMode: FlexSurfaceMode.levelSurfacesLowScaffold,
-//   blendLevel: 13,
-//   subThemesData: const FlexSubThemesData(
-//     blendOnLevel: 20,
-//     useTextTheme: true,
-//     useM2StyleDividerInM3: true,
-//     alignedDropdown: true,
-//     useInputDecoratorThemeInDialogs: true,
-//   ),
 //   visualDensity: FlexColorScheme.comfortablePlatformDensity,
 //   useMaterial3: true,
 //   swapLegacyOnMaterial3: true,
@@ -110,14 +156,6 @@
 //     error: Color(0xffcf6679),
 //   ),
 //   surfaceMode: FlexSurfaceMode.levelSurfacesLowScaffold,
-//   blendLevel: 13,
-//   subThemesData: const FlexSubThemesData(
-//     blendOnLevel: 20,
-//     useTextTheme: true,
-//     useM2StyleDividerInM3: true,
-//     alignedDropdown: true,
-//     useInputDecoratorThemeInDialogs: true,
-//   ),
 //   visualDensity: FlexColorScheme.comfortablePlatformDensity,
 //   useMaterial3: true,
 //   swapLegacyOnMaterial3: true,
@@ -135,10 +173,38 @@ import 'package:google_fonts/google_fonts.dart';
 class ThemeCollections {
   ThemeCollections._(); // Private constructor to prevent instantiation
 
-  static const List<String> availableThemes = ['sixPM', "inTheForest"];
+  static const List<String> availableThemes = [
+    'original',
+    'seaTheSky',
+    "inTheForest"
+  ];
   static final List<List> themeMapping = [
-    [sixPMThemeLight, sixPMThemeDark],
+    [originalThemeLight, originalThemeDark],
+    [seaTheSkyThemeLight, seaTheSkyThemeDark],
     [inTheForestThemeLight, inTheForestThemeDark]
+  ];
+
+  // Define common seaTheSky, inTheForest, and LukeCreated Original themes settings
+  static const _surfaceMode = FlexSurfaceMode.levelSurfacesLowScaffold;
+  static const _blendLevel = 7;
+  static const _subThemesData = FlexSubThemesData(
+    blendOnLevel: 10,
+    blendOnColors: false,
+    useTextTheme: true,
+    useM2StyleDividerInM3: true,
+    alignedDropdown: true,
+    useInputDecoratorThemeInDialogs: true,
+  );
+  static final _visualDensity = FlexColorScheme.comfortablePlatformDensity;
+  static const _useMaterial3 = true;
+  static const _swapLegacyOnMaterial3 = true;
+
+  // Define font families for all themes
+  static final _fontFamily = GoogleFonts.chewy().fontFamily;
+  static final _fontFamilyFallback = <String>[
+    GoogleFonts.krub().fontFamily!,
+    GoogleFonts.maShanZheng().fontFamily!,
+    GoogleFonts.notoSans().fontFamily!
   ];
 
   // Helper method to create ThemeData
@@ -176,18 +242,30 @@ class ThemeCollections {
     }
 
     availableThemes.elementAt(0);
-    return isDark ? sixPMThemeDark : sixPMThemeLight;
+    return isDark ? seaTheSkyThemeDark : seaTheSkyThemeLight;
   }
 
-  // Create sixPM Light theme data
-  static final sixPMThemeLight = _createThemeData(
-    colors: _sixPMLightColors,
+  // Create LukeCreated Original Light theme data
+  static final originalThemeLight = _createThemeData(
+    colors: _originalLightColors,
     brightness: Brightness.light,
   );
 
-  // Create sixPM Dark theme data
-  static final sixPMThemeDark = _createThemeData(
-    colors: _sixPMDarkColors,
+  // Create LukeCreated Original Dark theme data
+  static final originalThemeDark = _createThemeData(
+    colors: _originalDarkColors,
+    brightness: Brightness.dark,
+  );
+
+  // Create seaTheSky Light theme data
+  static final seaTheSkyThemeLight = _createThemeData(
+    colors: _seaTheSkyLightColors,
+    brightness: Brightness.light,
+  );
+
+  // Create seaTheSky Dark theme data
+  static final seaTheSkyThemeDark = _createThemeData(
+    colors: _seaTheSkyDarkColors,
     brightness: Brightness.dark,
   );
 
@@ -203,8 +281,32 @@ class ThemeCollections {
     brightness: Brightness.dark,
   );
 
-  // Define color schemes for sixPM Light Theme
-  static const FlexSchemeColor _sixPMLightColors = FlexSchemeColor(
+// Define color schemes for LukeCreated Original Light Theme
+  static const FlexSchemeColor _originalLightColors = FlexSchemeColor(
+    primary: Color(0xff1565c0),
+    primaryContainer: Color(0xff80d8ff),
+    secondary: Color(0xffffca28),
+    secondaryContainer: Color(0xffe4eaff),
+    tertiary: Color(0xffea80fc),
+    tertiaryContainer: Color(0xffcbd6ff),
+    appBarColor: Color(0xffe4eaff),
+    error: Color(0xffb00020),
+  );
+
+// Define color schemes for LukeCreated Original Dark Theme
+  static const FlexSchemeColor _originalDarkColors = FlexSchemeColor(
+    primary: Color(0xff00bcd4),
+    primaryContainer: Color(0xff3f51b5),
+    secondary: Color(0xffd6a400),
+    secondaryContainer: Color(0xffc9d8f2),
+    tertiary: Color(0xffc51162),
+    tertiaryContainer: Color(0xFF4C74CF),
+    appBarColor: Color(0xffc9d8f2),
+    error: Color(0xffcf6679),
+  );
+
+  // Define color schemes for seaTheSky Light Theme
+  static const FlexSchemeColor _seaTheSkyLightColors = FlexSchemeColor(
     primary: Color(0xffe65100),
     primaryContainer: Color(0xffffcc80),
     secondary: Color(0xff2979ff),
@@ -215,8 +317,8 @@ class ThemeCollections {
     error: Color(0xffb00020),
   );
 
-  // Define color schemes for sixPM Dark Theme
-  static const FlexSchemeColor _sixPMDarkColors = FlexSchemeColor(
+  // Define color schemes for seaTheSky Dark Theme
+  static const FlexSchemeColor _seaTheSkyDarkColors = FlexSchemeColor(
     primary: Color(0xffffb300),
     primaryContainer: Color(0xffc87200),
     secondary: Color(0xff82b1ff),
@@ -226,29 +328,6 @@ class ThemeCollections {
     appBarColor: Color(0xff3770cf),
     error: Color(0xffcf6679),
   );
-
-  // Define common sixPM, inTheForest themes settings
-  static const _surfaceMode = FlexSurfaceMode.levelSurfacesLowScaffold;
-  static const _blendLevel = 7;
-  static const _subThemesData = FlexSubThemesData(
-    blendOnLevel: 10,
-    blendOnColors: false,
-    useTextTheme: true,
-    useM2StyleDividerInM3: true,
-    alignedDropdown: true,
-    useInputDecoratorThemeInDialogs: true,
-  );
-  static final _visualDensity = FlexColorScheme.comfortablePlatformDensity;
-  static const _useMaterial3 = true;
-  static const _swapLegacyOnMaterial3 = true;
-
-  // Define font families for all themes
-  static final _fontFamily = GoogleFonts.chewy().fontFamily;
-  static final _fontFamilyFallback = <String>[
-    GoogleFonts.krub().fontFamily!,
-    GoogleFonts.maShanZheng().fontFamily!,
-    GoogleFonts.notoSans().fontFamily!
-  ];
 
   // Define color schemes for inTheForest Light Theme
   static const FlexSchemeColor _inTheForestLightColors = FlexSchemeColor(
@@ -262,6 +341,7 @@ class ThemeCollections {
     error: Color(0xffb00020),
   );
 
+  // Define color schemes for inTheForest Dark Theme
   static const FlexSchemeColor _inTheForestDarkColors = FlexSchemeColor(
     primary: Color(0xff7edb7b),
     primaryContainer: Color(0xff005313),